Add info about editing (#4)
* add info about editing

* remove unused file

* change header

* replace with float

* README auto update [skip ci]

---------

+++ CS/BlazorGridUnboundColumns/Data/WeatherForecast.cs
@@ -2,6 +2,6 @@ namespace BlazorGridUnboundColumns.Data {
     public class WeatherForecast {
         public DateTime Date { get; set; }
 
-        public int TemperatureC { get; set; }
+        public float TemperatureC { get; set; }
     }
 }
